[MJ-268] delete never verified accounts after 14 days

+++ registration/models.py
@@ -50,10 +50,10 @@ class RegistrationManager(models.Manager):
             try:
                 if profile.activation_key_expired():
                     user = profile.user
-                    if not user.is_active:
+                    if not user.is_disabled and not user.is_active:
                         user.delete()
                         profile.delete()
-            except User.DoesNotExist:
+            except DoesNotExist:
                 profile.delete()
